AB#62398 Possible fix for null depositService: Part 2

+++ backend/src/main/java/uk/ac/ox/it/calendarimporter/jobs/LoggingJob.java
@@ -132,9 +132,15 @@ public abstract class LoggingJob implements Job {
         result.problems = true;
     }
     
-    public void reset() {
+    public void reset() throws JobExecutionException {
         String logfile = progressService.resetJob(triggerId);
         if (logfile != null) {
+            if (depositService==null){
+                depositService = beanFactory.getBean("depositService", DepositService.class);
+            }
+            if (depositService==null){
+                throw new JobExecutionException("Failed to remove log file: " + logfile);
+            }
             depositService.remove(logfile);
         }
     }
